refactor(domain/board): 점수 계산의 기능을 ScoreCalculator 로 이전
