+++ src/main/java/chess/domain/gamestate/Running.java
@@ -35,12 +35,12 @@ public class Running implements State {
 
 	@Override
 	public double statusOfBlack() {
-		return board.calculateScoreOfBlack();
+		return board.scoreOfBlack();
 	}
 
 	@Override
 	public double statusOfWhite() {
-		return board.calculateScoreOfWhite();
+		return board.scoreOfWhite();
 	}
 
 	@Override
